refactor movie person

+++ source/src/main/java/com/tenant/api/controller/PersonController.java
@@ -4,12 +4,14 @@ import com.tenant.api.dto.ApiMessageDto;
 import com.tenant.api.dto.ErrorCode;
 import com.tenant.api.dto.ResponseListDto;
 import com.tenant.api.dto.person.PersonDto;
+import com.tenant.api.exception.BadRequestException;
 import com.tenant.api.exception.NotFoundException;
 import com.tenant.api.form.person.CreatePersonForm;
 import com.tenant.api.form.person.UpdatePersonForm;
 import com.tenant.api.mapper.PersonMapper;
 import com.tenant.api.storage.tenant.criteria.PersonCriteria;
 import com.tenant.api.storage.tenant.model.Person;
+import com.tenant.api.storage.tenant.repository.MoviePersonRepository;
 import com.tenant.api.storage.tenant.repository.PersonRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,6 +35,9 @@ public class PersonController extends ABasicController {
 
     @Autowired
     private PersonMapper personMapper;
+
+    @Autowired
+    private MoviePersonRepository moviePersonRepository;
 
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('PSN_C')")
@@ -87,7 +92,9 @@ public class PersonController extends ABasicController {
     public ApiMessageDto<Void> delete(@PathVariable("id") Long id) {
         Person person = personRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("[Person] Not found", ErrorCode.PERSON_ERROR_NOT_FOUND));
-
+        if (moviePersonRepository.existsByPersonId(person.getId())) {
+            throw new BadRequestException("[Person] Cannot delete with relationship with Movie Person", ErrorCode.PERSON_ERROR_MOVIE_PERSON_EXISTED);
+        }
         person.getKinds().clear();
         personRepository.delete(person);
         return makeSuccessResponse("Delete person success");
